PlantHolderVisualizer component system refactor (#10091)
* Converted PlantHolderVisualizer to component and system.

* Cleaned up Botany folder.

* Converted initialization of HydroTrayVisualizer under PlantHolderVisualizerSystem to it's own GenericVisualizer.

* Cleaned up hydro_tray.yml.
