+++ Content.Client/Botany/Components/PlantHolderVisualsComponent.cs
@@ -1,0 +1,6 @@
+namespace Content.Client.Botany.Components;
+
+[RegisterComponent]
+public sealed class PlantHolderVisualsComponent : Component
+{
+}
