+++ Content.Client/Botany/Components/PotencyVisualsComponent.cs
@@ -1,4 +1,4 @@
-﻿namespace Content.Client.Botany;
+namespace Content.Client.Botany.Components;
 
 [RegisterComponent]
 public sealed class PotencyVisualsComponent : Component
