+++ Content.Server/Entry/IgnoredComponents.cs
@@ -38,6 +38,7 @@ namespace Content.Server.Entry
             "PipeColorVisuals",
             "FireVisuals",
             "CrematoriumVisuals",
+            "PlantHolderVisuals",
         };
     }
 }
